# docs/swagger.yaml
basePath: /api/v1
definitions:
  echo.HTTPError:
    properties:
      message: {}
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  jsontype.Tags:
    additionalProperties: {}
    type: object
  omc.DataModel:
    properties:
      Created:
        type: string
      Default:
        type: boolean
      Deleted:
        $ref: '#/definitions/gorm.DeletedAt'
      Id:
        type: integer
      MetaData:
        $ref: '#/definitions/jsontype.Tags'
      Name:
        type: string
      ParameterPath:
        type: string
      Parameters:
        items:
          $ref: '#/definitions/omc.DataModelParameter'
        type: array
      ProductType:
        type: string
      Schema:
        type: string
      Templates:
        items:
          $ref: '#/definitions/omc.DataModelTemplate'
        type: array
      Updated:
        type: string
    type: object
  omc.DataModelParameter:
    properties:
      Created:
        type: string
      DataModelId:
        type: integer
      Default:
        type: boolean
      DefaultValue:
        type: string
      Deleted:
        $ref: '#/definitions/gorm.DeletedAt'
      Description:
        type: string
      Id:
        type: integer
      MetaData:
        $ref: '#/definitions/jsontype.Tags'
      Name:
        type: string
      Type:
        type: string
      Updated:
        type: string
      Writable:
        type: boolean
    type: object
  omc.DataModelTemplate:
    properties:
      Created:
        type: string
      DataModelId:
        type: integer
      Default:
        type: boolean
      Deleted:
        $ref: '#/definitions/gorm.DeletedAt'
      Id:
        type: integer
      MetaData:
        $ref: '#/definitions/jsontype.Tags'
      Name:
        type: string
      ParameterNames:
        $ref: '#/definitions/jsontype.Tags'
      Updated:
        type: string
    type: object
  omc.Device:
    properties:
      ActiveStatus:
        type: string
      Created:
        type: string
      Deleted:
        $ref: '#/definitions/gorm.DeletedAt'
      Enable:
        type: boolean
      Group:
        $ref: '#/definitions/omc.Group'
      GroupId:
        type: integer
      Id:
        type: integer
      LastInformTime:
        type: string
      MetaData:
        $ref: '#/definitions/jsontype.Tags'
      Methods:
        items:
          type: string
        type: array
      Name:
        type: string
      Online:
        type: boolean
      Oui:
        type: string
      ParameterNotifications:
        $ref: '#/definitions/omc.ParameterNotifications'
      ParameterValues:
        $ref: '#/definitions/omc.ParameterValues'
      ParameterWritables:
        $ref: '#/definitions/omc.ParameterWritables'
      Product:
        $ref: '#/definitions/omc.Product'
      ProductClass:
        type: string
      ProductId:
        type: integer
      ProductType:
        type: string
      Properties:
        $ref: '#/definitions/jsontype.Tags'
      Schema:
        type: string
      SerialNumber:
        type: string
      Updated:
        type: string
    type: object
...
  omc.Firmware:
    properties:
      Created:
        type: string
      Deleted:
        $ref: '#/definitions/gorm.DeletedAt'
      Id:
        type: integer
      Name:
        type: string
      ProductType:
        type: string
      S3Object:
        $ref: '#/definitions/s3.S3Object'
      S3ObjectId:
        type: integer
      Schema:
        type: string
      Updated:
        type: string
      UploadTime:
        type: string
      Uploader:
        type: string
      Version:
        type: string
      products:
        items:
          $ref: '#/definitions/omc.Product'
        type: array
    type: object
  omc.Group:
    properties:
      Children:
        items:
          $ref: '#/definitions/omc.Group'
        type: array
      Created:
        type: string
      Id:
        type: integer
      MetaData:
        $ref: '#/definitions/jsontype.Tags'
      Name:
        type: string
      Parent:
        $ref: '#/definitions/omc.Group'
      ParentId:
        type: integer
      Schema:
        type: string
      Updated:
        type: string
    type: object
...
  omc.ParameterNotifications:
    additionalProperties:
      type: integer
    type: object
  omc.ParameterValues:
    additionalProperties:
      type: string
    type: object
  omc.ParameterWritables:
    additionalProperties:
      type: boolean
    type: object
  omc.Product:
    properties:
      Created:
        type: string
      Default:
        type: boolean
      Deleted:
        $ref: '#/definitions/gorm.DeletedAt'
      Enable:
        type: boolean
      Firmwares:
        items:
          $ref: '#/definitions/omc.Firmware'
        type: array
      Id:
        type: integer
      Manufacturer:
        type: string
      MetaData:
        $ref: '#/definitions/jsontype.Tags'
      ModelName:
        type: string
      Oui:
        type: string
      ParameterPath:
        type: string
      PerformanceValueDefines:
        $ref: '#/definitions/jsontype.Tags'
      ProductClass:
        type: string
      ProductType:
        type: string
      Schema:
        type: string
      SupportedAlarms:
        items:
          $ref: '#/definitions/omc.ProductSupportedAlarm'
        type: array
      Updated:
        type: string
    type: object
  omc.ProductSupportedAlarm:
    properties:
      AlarmIdentifier:
        type: string
      Created:
        type: string
      EventType:
        type: string
      Id:
        type: integer
      Oui:
        type: string
      PerceivedSeverity:
        type: string
      ProbableCause:
        type: string
      Product:
        $ref: '#/definitions/omc.Product'
      ProductClass:
        type: string
      ProductId:
        type: integer
      ProductType:
        type: string
      Schema:
        type: string
      SerialNumber:
        type: string
      SpecificProblem:
        type: string
      Updated:
        type: string
    type: object
...
  omc.createDataModelBody:
    properties:
      metaData:
        $ref: '#/definitions/jsontype.Tags'
      name:
        type: string
      parameterPath:
        type: string
      productType:
        type: string
    type: object
  omc.createDataModelParameterBody:
    properties:
      defaultValue:
        type: string
      description:
        type: string
      metaData:
        $ref: '#/definitions/jsontype.Tags'
      name:
        type: string
      type:
        type: string
      writable:
        type: boolean
    type: object
...
  omc.listDataModelData:
    properties:
      Data:
        items:
          $ref: '#/definitions/omc.DataModel'
        type: array
      Total:
        type: integer
    type: object
  omc.listDeletedDevicesData:
    properties:
      Data:
        items:
          $ref: '#/definitions/omc.Device'
        type: array
      Total:
        type: integer
    type: object
...
  s3.S3Object:
    properties:
      Bucket:
        type: string
      Created:
        type: string
      Deleted:
        $ref: '#/definitions/gorm.DeletedAt'
      DownloadUrl:
        type: string
      FileName:
        type: string
      FileSize:
        type: integer
      Id:
        type: integer
      Key:
        type: string
      MetaData:
        $ref: '#/definitions/s3.Tags'
      Schema:
        type: string
      SignMethod:
        type: string
      Signature:
        type: string
      Updated:
        type: string
    type: object
  s3.Tags:
    additionalProperties: {}
    type: object
...
host: dev.netdoop.com
info:
  contact: {}
  title: NetDoop API
  version: "1.0"
paths:
...
  /omc/datamodels:
    get:
      operationId: list-datamodels
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      - default: ""
        description: Sort order
        in: query
        name: order_by
        type: string
      - default: ""
        description: Query
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Total:
              description: Total number
              type: int
          schema:
            $ref: '#/definitions/omc.listDataModelData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: List data models
      tags:
      - OMC DataModel
    post:
      operationId: create-datamodel
      parameters:
      - description: DataModel
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/omc.createDataModelBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/omc.DataModel'
        "400":
          description: 'Bad Request: invalid input parameter'
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Create data model
      tags:
      - OMC DataModel
  /omc/datamodels/{id}:
    delete:
      operationId: delete-datamodel
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Delete data model
      tags:
      - OMC DataModel
    get:
      operationId: get-datamodel
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/omc.DataModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Get data model
      tags:
      - OMC DataModel
  /omc/datamodels/{id}/parameter/{parameter_id}:
    delete:
      operationId: delete-datamodel-parameter
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: Parameter ID
        in: path
        name: parameter_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Delete data model parameter
      tags:
      - OMC DataModel Parameter
  /omc/datamodels/{id}/parameters:
    get:
      consumes:
      - application/json
      operationId: list-datamodel-parameters
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/omc.ParameterValues'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: List the parameters of a datamodel
      tags:
      - OMC DataModel
    post:
      operationId: create-datamodel-parameter
      parameters:
      - description: DataModelParameter
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/omc.createDataModelParameterBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/omc.DataModelParameter'
        "400":
          description: 'Bad Request: invalid input parameter'
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Create data model parameter
      tags:
      - OMC DataModel Parameter
  /omc/datamodels/{id}/parameters/{parameter_id}:
    get:
      operationId: get-datamodel-parameter
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: Parameter ID
        in: path
        name: parameter_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/omc.DataModelParameter'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Get data model parameter
      tags:
      - OMC DataModel Parameter
...
  /omc/deleted-devices:
    get:
      consumes:
      - application/json
      operationId: list-deleted-devices
      parameters:
      - default: ""
        description: Query
        in: query
        name: q
        type: string
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      - default: ""
        description: Sort order
        in: query
        name: order_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/omc.listDeletedDevicesData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: List deleted devices
      tags:
      - OMC Deleted Devices
  /omc/deleted-devices/{id}:
    delete:
      consumes:
      - application/json
      operationId: delete-deleted-device
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Delete a device permanently
      tags:
      - OMC Deleted Devices
  /omc/deleted-devices/{id}/recover:
    post:
      consumes:
      - application/json
      operationId: recover-deleted-device
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Recover a deleted device
      tags:
      - OMC Deleted Devices
...
schemes:
- http
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"